requirements.txt
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
pydantic>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
sseclient>=0.0.27
httpx>=0.24.1
orjson>=3.9.0
email-validator>=2.0.0
tenacity>=8.2.2